+++ app/src/main/kotlin/de/codevoid/screensaver/BootReceiver.kt
@@ -1,0 +1,13 @@
+package de.codevoid.screensaver
+
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+
+class BootReceiver : BroadcastReceiver() {
+    override fun onReceive(context: Context, intent: Intent) {
+        if (intent.action == Intent.ACTION_BOOT_COMPLETED) {
+            context.startForegroundService(Intent(context, BrightnessService::class.java))
+        }
+    }
+}